refactor(monitoring): split reserve verification modules
